junit: Import codeowners via test case properties
Extend the testcase format to read properties in general, and the owner
property in particular, from junit files generated by the Cilium test
frameworks. This will enable us to track the owners for both passing and
failing tests, which we can later filter on to determine failure rates
per code owner.

Signed-off-by: Joe Stringer <joe@isovalent.com>