junit: Do not require owners data to include test name
So far we've been expecting the owners data to include a test name or
reference like "@cilium/sig-agent (test-name)", but the test name is
already in the parent object. This commit loosens the requirement so
that other newer junit formats like the one to be generated as part of
unit testing do not need to provide the same form.

Signed-off-by: Joe Stringer <joe@isovalent.com>